refactor: move hyperopt-output
